# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trada AI trading platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/deposit-address:
    get:
      operationId: getDepositAddress
      tags: [wallet]
      summary: Get deposit address for a network
      parameters:
        - name: network
          in: query
          required: true
          schema:
            type: string
            enum: [ETH, BTC, USDT]
      responses:
        "200":
          description: Deposit address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositAddress"
  /wallet/transactions:
    get:
      operationId: listTransactions
      tags: [wallet]
      summary: List transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, withdrawal, trade_fee, rental_fee, adjustment]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /wallet/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [wallet]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequestBody"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
        "400":
          description: Insufficient balance or invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── AGENTS ───────────────────────────────────────────────────────────────
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List user's trade history
      parameters:
        - name: rentalId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, executed, failed, skipped]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trade list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeList"
  /trades/{tradeId}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get trade details including AI signal
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trade details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    DepositAddress:
      type: object
      properties:
        address:
          type: string
        network:
          type: string
        qrCode:
          type: string
          nullable: true
        memo:
          type: string
          nullable: true
      required: [address, network]
    Transaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [deposit, withdrawal, trade_fee, rental_fee, adjustment]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        description:
          type: string
        txHash:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, type, amount, currency, status, description, createdAt]
    TransactionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required: [items, total, offset, limit]
    WithdrawalRequestBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        address:
          type: string
        network:
          type: string
          enum: [ETH, BTC, USDT]
        memo:
          type: string
      required: [amount, address, network]
    WithdrawalRequest:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        amount:
          type: number
        currency:
          type: string
        address:
          type: string
        network:
          type: string
        status:
          type: string
          enum: [requested, approved, broadcasted, completed, failed, rejected]
        rejectionReason:
          type: string
          nullable: true
        txHash:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, amount, currency, address, network, status, createdAt, updatedAt]
...
    Trade:
      type: object
      properties:
        id:
          type: string
        rentalId:
          type: string
        agentId:
          type: string
        agentName:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell, hold]
        status:
          type: string
          enum: [pending, executed, failed, skipped]
        entryPrice:
          type: number
          nullable: true
        exitPrice:
          type: number
          nullable: true
        quantity:
          type: number
          nullable: true
        pnl:
          type: number
          nullable: true
        signalConfidence:
          type: number
          nullable: true
        failureReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - rentalId
        - agentId
        - agentName
        - symbol
        - side
        - status
        - createdAt
    TradeDetail:
      allOf:
        - $ref: "#/components/schemas/Trade"
        - type: object
          properties:
            signalRationale:
              type: string
              nullable: true
            rawSignal:
              type: object
              nullable: true
              additionalProperties: true
            exchangeResponse:
              type: object
              nullable: true
              additionalProperties: true
          required: []
    TradeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required: [items, total, offset, limit]

    # ── NOTIFICATIONS ───────────────────────────────────────────────────────
...