# lib/api-zod/src/generated/types/withdrawalRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trada AI trading platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WithdrawalRequestBodyNetwork } from "./withdrawalRequestBodyNetwork";

export interface WithdrawalRequestBody {
  /** @minimum 0.01 */
  amount: number;
  address: string;
  network: WithdrawalRequestBodyNetwork;
  memo?: string;
}